can add a blank group blocks to data base
