+++ client/lib/fillGroupBlocks.js
@@ -1,0 +1,51 @@
+
+function fillGroupBlocks(userId, userBlocks, groupBlocks) {
+  const fillBlocks = groupBlocks.blocks;
+  console.log(userBlocks.selected);
+  console.log(fillBlocks);
+
+  for (let i = 0; i < userBlocks.selected.length; i++) {
+    let { row, col } = userBlocks.selected[i];
+    row = parseInt(row);
+    col = parseInt(col);
+    fillBlocks[row][col].push(userId);
+  }
+  return fillBlocks;
+}
+
+
+// export default fillGroupBlocks;
+
+const userId = 14;
+
+const userBlocks = {
+  selected: [
+    {
+      row: '0',
+      col: '0'
+    },
+    {
+      row: '0',
+      col: '1'
+    },
+    {
+      row: '0',
+      col: '2'
+    },
+    {
+      row: '0',
+      col: '3'
+    }
+  ]
+}
+
+const groupBlocks ={
+  blocks: [
+    [[], [], [], [], []],
+    [[], [], [], [], []],
+    [[], [], [], [], []],
+    [[], [], [], [], []]
+  ]
+}
+
+console.log(fillGroupBlocks(userId, userBlocks, groupBlocks));
