+++ client/pages/meeting-event.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import jwtDecode from 'jwt-decode';
 import Header from '../components/header.jsx';
 import Button from '../components/button.jsx';
+import UserMeetingBlocks from '../components/user-meeting-blocks.jsx';
 import returnTimesArr from '../lib/returnTimesArr.js';
 import fillSelectSquare from '../lib/fillSelectSquare.js';
 const some = require('lodash/some');
@@ -13,7 +14,7 @@ function MeetingTitle(props) {
       <h1 className='font-nunito-sans text-3xl font-thin mt-8'>
         {props.name}
       </h1>
-      <p className='font-nunito-sans text-sm font-thin text-gray-500 mt-5 mb-4'>
+      <p className='font-nunito-sans text-sm font-thin text-gray-500 mt-5 mb-10'>
         {props.description}
       </p>
     </div>
@@ -51,6 +52,8 @@ function RegistrationForm(props) {
 }
 
 class GroupMeetingBlocks extends React.Component {
+
+  // on component did mount load the selected blocks or fetch it
 
   render() {
     const { dates, startTime, endTime } = this.props;
@@ -125,238 +128,6 @@ class GroupMeetingBlocks extends React.Component {
           {blocks}
         </div>
       </div>
-      </>
-    );
-  }
-}
-
-class UserMeetingBlocks extends React.Component {
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      blocks: {
-        selected: []
-      },
-      selecting: false,
-      toggle: false,
-      fill: {
-        init: null
-      },
-      isAuthorizing: true
-    };
-
-    this.handleMouseDown = this.handleMouseDown.bind(this);
-    this.handleMouseUp = this.handleMouseUp.bind(this);
-    this.handleMouseEnter = this.handleMouseEnter.bind(this);
-  }
-
-  componentDidMount() {
-    const { route, user } = this.props;
-    const meetingId = route.params.get('meetingId');
-    const userId = user.userId;
-
-    fetch(`/api/users/${userId}/meetingId/${meetingId}`)
-      .then(res => res.json())
-      .then(user => {
-        const { selectedTimes } = user;
-        this.setState({
-          blocks: selectedTimes
-        });
-      })
-      .catch(err => console.error(err));
-
-    this.setState({ isAuthorizing: false });
-  }
-
-  componentDidUpdate() {
-    const { route, user } = this.props;
-    const { blocks } = this.state;
-    const meetingId = route.params.get('meetingId');
-    const userId = user.userId;
-
-    const body = { meetingId, blocks };
-    const req = {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify(body)
-    };
-
-    fetch(`/api/users/${userId}`, req)
-      .then(res => res.json())
-      .catch(err => console.error(err));
-  }
-
-  handleMouseDown(event) {
-    this.setState({
-      toggle: true
-    });
-
-    const col = event.target.getAttribute('col');
-    const row = event.target.getAttribute('row');
-    const block = { col, row };
-    const { selected } = this.state.blocks;
-
-    this.setState(prevState => ({
-      fill: {
-        ...prevState.fill,
-        init: block
-      }
-    }));
-
-    if (!some(selected, block)) {
-      this.setState({
-        blocks: {
-          selected: [...this.state.blocks.selected, block]
-        },
-        selecting: true
-      });
-    } else {
-      const removeIndex = selected.findIndex(val => {
-        return val.col === block.col && val.row === block.row;
-      });
-      selected.splice(removeIndex, 1);
-      this.setState({
-        blocks: {
-          selected
-        },
-        selecting: false
-      });
-    }
-  }
-
-  handleMouseUp(event) {
-    this.setState({ toggle: false });
-  }
-
-  handleMouseEnter(event) {
-    if (!this.state.toggle) return;
-
-    const col = event.target.getAttribute('col');
-    const row = event.target.getAttribute('row');
-    const block = { col, row };
-    const { selected } = this.state.blocks;
-
-    if (this.state.selecting && some(selected, block)) return;
-
-    const { init } = this.state.fill;
-    let fill = [];
-    if (this.state.selecting) {
-
-      fill = fillSelectSquare(init, block).filter(val => !some(selected, val));
-      this.setState({
-        blocks: {
-          selected: [...this.state.blocks.selected, ...fill]
-        }
-      });
-    } else if (!this.state.selecting) {
-      if (!some(selected, block)) return;
-
-      fill = fillSelectSquare(init, block);
-      const removed = selected.filter(val => !some(fill, val));
-
-      this.setState({
-        blocks: {
-          selected: removed
-        }
-      });
-    }
-  }
-
-  render() {
-    if (!this.state.blocks.selected) return null;
-
-    const { handleMouseDown, handleMouseUp, handleMouseEnter } = this;
-
-    const { selected } = this.state.blocks;
-
-    const { dates, startTime, endTime } = this.props;
-    if (dates.length === 0) return;
-    const days = JSON.parse(dates).days;
-    const times = returnTimesArr(startTime, endTime);
-
-    const rows = [];
-    for (let i = 0; i < times.length - 1; i++) {
-
-      const row = [];
-      for (let j = 0; j < days.length; j++) {
-
-        let color = 'bg-gray-300';
-        for (let k = 0; k < selected.length; k++) {
-
-          if (selected[k].row === i.toString() && selected[k].col === j.toString()) {
-            color = 'bg-green-500';
-            break;
-          }
-
-        }
-
-        const block = (
-          <div key={j} time={times[i]} date={days[j]} col={j} row={i} onMouseDown={handleMouseDown} onMouseUp={handleMouseUp} onMouseEnter={handleMouseEnter}
-            className={`h-3 w-14 ${color} mr-0.5`}></div>
-        );
-        row.push(block);
-
-      }
-
-      rows.push(row);
-    }
-
-    const blocks = rows.map((row, index) =>
-      <div key={index} className='blocks flex justify-center'>
-        {row}
-      </div>
-    );
-
-    const hours = times.filter(time => {
-      for (let i = 0; i < time.length; i++) {
-        if (time[i] === ':') {
-          if (time[i + 1] === '0' && time[i + 2] === '0') {
-            return true;
-          }
-        }
-      }
-      return false;
-    });
-
-    const timeLabels = hours.map(hour => {
-      return (
-        <div key={hour} className='font-plus-jakarta-sans text-xxs font-light'>
-          {hour}
-        </div>
-      );
-    });
-
-    const dateLabels = days.map(day => {
-      const dayArr = day.split(' ');
-
-      return (
-        <div key={day} className='w-14 text-center font-plus-jakarta-sans'>
-          <div className='text-xs'>
-            {[dayArr[1], dayArr[2]].join(' ')}
-          </div>
-          <div className='text-base'>
-            {dayArr[0][0]}
-          </div>
-        </div>
-      );
-    });
-
-    return (
-      <>
-        <div className='flex justify-center overflow-x-scroll select-none'>
-          <div className='flex flex-col justify-between mt-20 mb-10 mr-1'>
-            {timeLabels}
-          </div>
-          <div className='my-10 w-min flex flex-wrap justify-center'>
-            <div className='w-full flex justify-around'>
-              {dateLabels}
-            </div>
-            {blocks}
-          </div>
-        </div>
       </>
     );
   }
@@ -457,7 +228,7 @@ export default class MeetingEvent extends React.Component {
   render() {
     if (this.state.isAuthorizing) return null;
 
-    const { name, description, dates, startTime, endTime, user } = this.state;
+    const { name, description, dates, startTime, endTime, user, selectedBlocks } = this.state;
     const { handleRegistration, handleUserName, handleSignOut } = this;
 
     const signOut = user
@@ -466,7 +237,7 @@ export default class MeetingEvent extends React.Component {
       : null;
 
     const userView = user
-      ? <UserMeetingBlocks dates={dates} startTime={startTime} endTime={endTime} route={this.props.route} user={user} />
+      ? <UserMeetingBlocks dates={dates} startTime={startTime} endTime={endTime} route={this.props.route} user={user} groupBlocks={selectedBlocks} />
       : <RegistrationForm handleUserName={handleUserName} handleRegistration={handleRegistration} label="Register as a participant" />;
     return (
       <>
@@ -475,8 +246,8 @@ export default class MeetingEvent extends React.Component {
                         lg:w-4/5 lg:m-auto lg:mt-5'>
           <div className='w-96 lg:w-1/2 lg:h-116'>
             <MeetingTitle name={name} description={description} />
-            {signOut}
             {userView}
+            {signOut}
           </div>
           <div className='w-full lg:w-1/2 text-center flex flex-wrap justify-center'>
             <h1 className='font-nunito-sans text-xl font-thin mt-10 w-full'>
@@ -485,7 +256,7 @@ export default class MeetingEvent extends React.Component {
             <h3 className='font-nunito-sans mt-4 mb-10 w-full'>
               0 registered
             </h3>
-            <GroupMeetingBlocks dates={dates} startTime={startTime} endTime={endTime} />
+            <GroupMeetingBlocks dates={dates} startTime={startTime} endTime={endTime} groupBlocks={selectedBlocks}/>
           </div>
         </div>
       </>
